+++ src/main/java/app/domain/shared/ListingPriceComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class ListingPriceComparator implements Comparator<Listing> {
     @Override
     public int compare(Listing o1, Listing o2) {
-        return Integer.compare(o1.getPrice(), o2.getPrice());
+        return Float.compare(o1.getPrice(), o2.getPrice());
     }
 }
